RecoNet/requirements.txt
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
email_validator==2.2.0
Werkzeug==3.1.3
WTForms==3.2.1
python-dotenv==1.0.1

# WHOIS
python-whois==0.9.5

# IP WHOIS
ipwhois==1.3.0

# DNS Lookup
dnspython==2.7.0

# Geolocation
geopy==2.4.1
